Create button border trialing (failed)
Basically I was making an attempt to make the border of the buttons to black. It failed, I was going to try and use tk.frame but in order to use this method I would have to change other methods of my code to adjust to the import tk frame. So i will just not have a black border for my button.