Create second image is working, used different method
I couldn't get the image to appear on the new windows before but I managed to fix that problem by changing my coding method. Instead of using a new image for the new window or reopening the first one, I used the previously stated variable name of the image on the opening window.